feat #1: 1주차 코드 정리
